Added crew member attributes to show page
The show page now displays the crew for the movie (currently chadwick
boseman is a producer on Star Wars 9). As with the cast, clicking his
name takes you to his person page. Right now cast and crew are
differentiated by checking whether their character role is null, this
may need to be changed but this is functioning code